+++ zarrs/src/array/array_ops/async_array_read_ops.rs
@@ -159,6 +159,7 @@ pub trait AsyncArrayReadOps: ArrayOps {
     ) -> Result<T, ArrayError> {
         let subchunk_grid = self
             .subchunk_grid_at_level(level)
+            .as_chunk_grid()
             .ok_or(ArrayError::MissingSubchunkGrid)?;
         let array_subset = subchunk_grid
             .subset(subchunk_indices)?
@@ -188,6 +189,7 @@ pub trait AsyncArrayReadOps: ArrayOps {
     ) -> Result<T, ArrayError> {
         let subchunk_grid = self
             .subchunk_grid_at_level(level)
+            .as_chunk_grid()
             .ok_or(ArrayError::MissingSubchunkGrid)?;
         let array_subset = subchunk_grid.chunks_subset(subchunks)?.ok_or_else(|| {
             ArrayError::InvalidArraySubset(